create setNotificationInviteDeclined function
