+++ frontend/src/util/middleware/socketActions/notification.tsx
@@ -1,7 +1,8 @@
 /**
  * Socket callbacks for setting notification
  */
-import { setNotification } from "../../../app/features/notificationSlice";
+import { setNotification, resetNotification } from "../../../app/features/notificationSlice";
+import { User } from "../../../app/features/types";
 import { store } from "../../../app/store";
 
 export const setNotificationSendInvite = (peerUsername: string) => {
@@ -13,4 +14,15 @@ export const setNotificationSendInvite = (peerUsername: string) => {
   }
 
   store.dispatch(setNotification(notificationData))
+}
+
+export const setNotificatioInviteDeclined = (peer: User) => {
+  const notificationData = {
+    notificationContent: `${peer.username} is not able to chat`,
+    notificationType: 'is-warning',
+    isLoading: false,
+    isActive: true,
+  }
+  store.dispatch(setNotification(notificationData))
+  setTimeout(() => store.dispatch(resetNotification()), 5000)
 }
